fix: wire submodules to shared resolver state
- Connect _query_resolve and _url_resolve to facade's _circuit_breakers and _routing_memory
- Add regression test to ensure submodules share state with facade
- Ensures consistent runtime behavior and test compatibility
